third_party/blink/web_tests/editing/pasteboard/paste-text-with-style-2-expected.txt

This tests copying and pasting text does not result in nested style elements such as b, em, and font.
To manually test, copy and paste "hello world" below. WebKit should not nest b, i, u, and font.

Before copy-paste:
| <b>
|   <i>
|     <u>
|       <font>
|         color="red"
|         "<#selection-caret>hello world"

After copy-paste:
| <b>
|   <i>
|     <u>
|       <font>
|         color="red"
|         "hello world<#selection-caret>"
